It is finally fully functional
Reduced mouse lag and issues.
Flying now somewhat moves you in the camera direction.

+++ C-Craft++/World/Generation/Biome/Biome.cpp
@@ -1,29 +1,31 @@
 #include "Biome.h"
 
 Biome::Biome(const NoiseParameters& parameters, int treeFreq, int plantFreq, int seed)
-	: m_heightGenerator(seed)
-	, m_treeFreq (treeFreq)
-	, m_plantFreq(plantFreq)
+:   m_heightGenerator   (seed)
+,   m_treeFreq      (treeFreq)
+,   m_plantFreq     (plantFreq)
 {
-	m_heightGenerator.setParameters(parameters);
+    m_heightGenerator.setParameters(parameters);
 }
 
-ChunkBlock Biome::getBeachBlock(Rand& rand)const
+ChunkBlock Biome::getBeachBlock(Rand& rand) const
 {
-	return BlockID::Sand;
+    return BlockID::Sand;
 }
+
 
 int Biome::getHeight(int x, int z, int chunkX, int chunkZ) const
 {
-	return (int)m_heightGenerator.getHeight(x, z, chunkX, chunkZ);
+    return m_heightGenerator.getHeight(x, z, chunkX, chunkZ);
 }
 
 int Biome::getTreeFrequency() const noexcept
 {
-	return m_treeFreq;
+    return m_treeFreq;
 }
 
-int Biome::getPlantFrequency() const noexcept
+int Biome::getPlantFrequency () const noexcept
 {
-	return m_plantFreq;
+    return m_plantFreq;
 }
+
